beaucoup de chose jeudi

+++ src/acelerometre.h
@@ -2,6 +2,6 @@
 #define ACCELEROMETRE_H
 
 int setup_accelerometre();
-float module_accelerometre();
+int module_accelerometre();
 
 #endif
